polyglot_backend/flask_service/requirements.txt
Flask==3.1.1
Flask-Cors==6.0.1
Flask-SocketIO==5.5.1
python-dotenv==1.1.1
requests==2.32.4
eventlet==0.40.3